Conflicts:
	arch/arm/boot/dts/qcom/msm8917-camera.dtsi
	arch/arm/boot/dts/qcom/msm8937-camera.dtsi
	arch/arm/boot/dts/qcom/msm8953-camera.dtsi

+++ arch/arm/mm/dma-mapping.c
@@ -1211,8 +1211,8 @@ static struct page **__iommu_alloc_buffer(struct device *dev, size_t size,
 					  gfp_t gfp, struct dma_attrs *attrs)
 {
 	struct page **pages;
-	int count = size >> PAGE_SHIFT;
-	int array_size = count * sizeof(struct page *);
+	size_t count = size >> PAGE_SHIFT;
+	size_t array_size = count * sizeof(struct page *);
 	int i = 0;
 
 	if (array_size <= PAGE_SIZE)
